feat: fix img, add new select
