+++ frontend/components/playground/playground.tsx
@@ -6,7 +6,7 @@ import { useCallback, useEffect, useState } from "react";
 import { Controller, FormProvider, SubmitHandler, useForm } from "react-hook-form";
 
 import Messages from "@/components/playground/messages";
-import LlmSelectNew from "@/components/playground/messages/llm-select-new";
+import LlmSelect from "@/components/playground/messages/llm-select";
 import { useToast } from "@/lib/hooks/use-toast";
 import { Message, Playground as PlaygroundType, PlaygroundForm } from "@/lib/playground/types";
 import { mapMessages, parseSystemMessages, remapMessages } from "@/lib/playground/utils";
@@ -16,7 +16,6 @@ import { Button } from "../ui/button";
 import Formatter from "../ui/formatter";
 import Header from "../ui/header";
 import { ScrollArea } from "../ui/scroll-area";
-import LlmSelect from "./messages/llm-select";
 
 const defaultMessages: Message[] = [
   {
@@ -34,9 +33,9 @@ const renderMessages = (messages: Message[], inputs: Record<string, string>): Me
     content: message.content.map((content) =>
       content.type === "text"
         ? {
-            ...content,
-            text: renderText(content.text, inputs),
-          }
+          ...content,
+          text: renderText(content.text, inputs),
+        }
         : content
     ),
   }));
@@ -148,7 +147,6 @@ export default function Playground({ playground }: { playground: PlaygroundType 
           <div className="flex flex-col gap-4 p-4">
             <div className="flex flex-col gap-2"></div>
             <FormProvider {...methods}>
-              <LlmSelectNew />
               <Controller
                 render={({ field: { value, onChange } }) => <LlmSelect value={value} onChange={onChange} />}
                 name="model"
